The trait TestResult is insufficient. Can't handle the Box<OFAst> case where the type is always Box<OFAst> but the checks are always different. So this trait is dead now.
Have a simpler on TestEq which mimics PartialEq. (see Trait impl restrictions). This still doesn't work for Box<OFAst>, so I've given up on that one. Have to use okeq() for this case, which is good enough.

Replaced the rest of TestResult with a sprinkling of Debug and Display. Doesn't give such a nice output, but we're not finished.

Add more tests...